sound_scaffold/audio_analyzer.py
import os
import json
import numpy as np
import librosa
import tensorflow as tf
from google.cloud import storage
from google.cloud import firestore

class AudioAnalyzer:
    """Audio analysis engine for SoundScaffold.
    
    This class provides functionality to analyze audio files and extract scene information,
    transitions, and quality metrics.
    """
    
    DEFAULT_CONFIG = {
        'sample_rate': 22050,
        'n_fft': 2048,
        'hop_length': 512,
        'n_mels': 128,
        'model_path': 'models/audio_analyzer_model.h5',
        'scene_threshold': 0.65,
        'quality_threshold': 0.5
    }
    
    def __init__(self, config=None):
        """Initialize the AudioAnalyzer with optional custom configuration.
        
        Args:
            config (dict, optional): Configuration dictionary. Defaults to None.
        """
        self.config = config or self.DEFAULT_CONFIG
        self.model = self._load_model()
        self.storage_client = storage.Client()
        self.db = firestore.Client()
        
    def _load_model(self):
        """Load the TensorFlow model for audio analysis.
        
        Returns:
            tf.keras.Model: The loaded model.
        """
        # In a real implementation, this would load a trained model
        # For this example, we'll create a simple dummy model
        try:
            if os.path.exists(self.config['model_path']):
                return tf.keras.models.load_model(self.config['model_path'])
            else:
                print(f"Model not found at {self.config['model_path']}. Using dummy model.")
                return self._create_dummy_model()
        except Exception as e:
            print(f"Error loading model: {e}. Using dummy model.")
            return self._create_dummy_model()
    
    def _create_dummy_model(self):
        """Create a dummy model for demonstration purposes.
        
        Returns:
            tf.keras.Model: A simple dummy model.
        """
        inputs = tf.keras.Input(shape=(self.config['n_mels'], None, 1))
        x = tf.keras.layers.Conv2D(16, (3, 3), activation='relu')(inputs)
        x = tf.keras.layers.GlobalAveragePooling2D()(x)
        outputs = tf.keras.layers.Dense(10, activation='softmax')(x)
        model = tf.keras.Model(inputs=inputs, outputs=outputs)
        return model
    
    def analyze(self, audio_file):
        """Analyze an audio file and return scene breakdown.
        
        Args:
            audio_file (str): Path to the audio file.
            
        Returns:
            dict: Analysis results containing scene breakdown, quality metrics, etc.
        """
        try:
            features = self._extract_features(audio_file)
            scene_data = self._classify_scenes(features)
            quality_metrics = self._assess_quality(features)
            transitions = self._detect_transitions(features)
            
            # Combine all analysis results
            results = {
                'file_name': os.path.basename(audio_file),
                'duration': features.get('duration', 0),
                'scene_breakdown': scene_data,
                'quality_metrics': quality_metrics,
                'transitions': transitions,
                'timestamp': firestore.SERVER_TIMESTAMP
            }
            
            # Store results in Firestore
            doc_ref = self.db.collection('audio_analyses').document()
            doc_ref.set(results)
            results['analysis_id'] = doc_ref.id
            
            return results
        except Exception as e:
            print(f"Error analyzing audio file: {e}")
            return {'error': str(e)}
    
    def _extract_features(self, audio_file):
        """Extract audio features from file.
        
        Args:
            audio_file (str): Path to the audio file.
            
        Returns:
            dict: Dictionary of extracted features.
        """
        try:
            # Load audio file
            y, sr = librosa.load(audio_file, sr=self.config['sample_rate'])
            
            # Extract duration
            duration = librosa.get_duration(y=y, sr=sr)
            
            # Compute mel spectrogram
            mel_spec = librosa.feature.melspectrogram(
                y=y,
                sr=sr,
                n_fft=self.config['n_fft'],
                hop_length=self.config['hop_length'],
                n_mels=self.config['n_mels']
            )
            log_mel_spec = librosa.power_to_db(mel_spec)
            
            # Extract MFCC features
            mfcc = librosa.feature.mfcc(
                S=librosa.power_to_db(mel_spec),
                n_mfcc=20
            )
            
            # Extract other useful features
            spectral_contrast = librosa.feature.spectral_contrast(y=y, sr=sr)
            chroma = librosa.feature.chroma_stft(y=y, sr=sr)
            onset_strength = librosa.onset.onset_strength(y=y, sr=sr)
            tempo, _ = librosa.beat.beat_track(onset_envelope=onset_strength, sr=sr)
            
            return {
                'duration': duration,
                'mel_spectrogram': log_mel_spec,
                'mfcc': mfcc,
                'spectral_contrast': spectral_contrast,
                'chroma': chroma,
                'tempo': tempo,
                'raw_audio': y,
                'sample_rate': sr
            }
        
        except Exception as e:
            print(f"Error extracting features: {e}")
            raise
    
    def _classify_scenes(self, features):
        """Classify audio scenes using the loaded model.
        
        Args:
            features (dict): Extracted audio features.
            
        Returns:
            list: Scene classifications with timestamps.
        """
        # In a real implementation, this would use the model to classify scenes
        # For this example, we'll generate mock scene data
        
        duration = features.get('duration', 0)
        segments = int(duration / 5)  # Segment into 5-second chunks
        scene_types = ['dialogue', 'ambient', 'music', 'action', 'silence']
        
        scenes = []
        current_time = 0
        
        for i in range(segments):
            segment_length = min(5, duration - current_time)
            scene_type = scene_types[i % len(scene_types)]
            confidence = 0.7 + (np.random.random() * 0.3)  # Random confidence between 0.7 and 1.0
            
            scenes.append({
                'start_time': current_time,
                'end_time': current_time + segment_length,
                'scene_type': scene_type,
                'confidence': float(confidence)
            })
            
            current_time += segment_length
        
        return scenes
    
    def _assess_quality(self, features):
        """Assess audio quality metrics.
        
        Args:
            features (dict): Extracted audio features.
            
        Returns:
            dict: Quality metrics.
        """
        # In a real implementation, this would compute actual quality metrics
        # For this example, we'll generate mock quality data
        
        return {
            'signal_to_noise': 45.3,  # dB
            'clarity': 0.82,  # 0-1 scale
            'distortion': 0.05,  # 0-1 scale (lower is better)
            'dynamic_range': 60.2,  # dB
            'suggested_improvements': [
                'Minor noise reduction recommended',
                'Consider light compression to improve dialogue clarity'
            ]
        }
    
    def _detect_transitions(self, features):
        """Detect audio transitions and scene changes.
        
        Args:
            features (dict): Extracted audio features.
            
        Returns:
            list: Detected transitions with timestamps.
        """
        # In a real implementation, this would detect actual transitions
        # For this example, we'll generate mock transition data
        
        duration = features.get('duration', 0)
        num_transitions = int(duration / 15)  # Roughly one transition every 15 seconds
        
        transitions = []
        transition_types = ['cut', 'fade', 'cross-fade', 'ambient-shift']
        
        for i in range(num_transitions):
            transition_time = (i + 1) * 15 + (np.random.random() * 5 - 2.5)  # Add some randomness
            if transition_time < duration:
                transitions.append({
                    'time': float(transition_time),
                    'type': transition_types[i % len(transition_types)],
                    'sharpness': float(0.4 + (np.random.random() * 0.6))  # 0-1 scale
                })
        
        return transitions
