requirements.txt
# Core dependencies
numpy>=1.22.0
librosa>=0.9.0
tensorflow>=2.10.0
soundfile>=0.10.0
flask>=2.0.0
gunicorn>=20.1.0

# Google Cloud dependencies
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.3.0
google-cloud-language>=2.0.0

# Additional utilities
werkzeug>=2.0.0
pydub>=0.25.0
